requirements.txt
numpy
pandas
scikit-learn
xgboost
SimpleITK
nibabel
pyradiomics
torch
torchvision
tqdm
pydicom
tcia-utils
